feat: implement 'Access-State' filter in backend
To filter on rights the query needs to be changed significantly.
With that in mind there is a query type for when rights are filtered
and if not. The query for filtering rights got implemented in this
commit. It uses a left outer join over three tables, beginning
with 'item_metadata' then 'item' and finally 'item_right'.
The where clause in that query takes care of all 'item_metadata' related
filters and after the join all metadata entries are related with each
right they have. Then the filters for rights can be applied.
Afterwards it is important to distinct on the metadata_id field.