fix: replace underscore enum values
When it comes to enum values with an underscore the generator generates
different cases for typescript and kotlin. Kotlin will get camel case
while typescript receives snake case. This caused me a lot of headache.
Since there doesn't seem to be an easy solution, the underscores are
removed and it is one big unreadable term for now.