feat: exceptions finished
Approach was changed in this last commit. Collecting all bookmarks and
their filters conjunction them and use their negation in the queries
turned out not to work. This is due to how queries are constructed.
Different filters are in different part of queries which made this
approach impossible. Now all exception queries are executed first, the
metadata_ids are collected and then the 'real' query is done with
an additional constraint of not allowing any results having one of the
metadata_ids of the exceptions.