feat: index series zdb
Apparently there is another zdb id in town - ZDB ID for series. The
requirement was to show all series zdb ids in the existing zdb id
filter. To keep it as clean as possible the existing zdb id was renamed
to journal zdb id and whenever a zdb id filter is set or a key in the
searchbar is used, the backend searches for both. Facets are adjusted
accordingly so that all journal and series ids are added together