feat: filter on template names no longer rely on name
Instead the right id is significant. The template name will be
shown in the ui and the filter `tpl` does still exist. But when using
the template filter in the left search bar the new filter `rightid`
comes to use. This allows saving bookmarks with this filter and changing
the name of template afterwards, which was the feature request here.