fix(636): big queries no longer fail & searching multiple handle is possible
Description:
Searching for `hdl:11159/632043 | hdl:11159/632044 | ...` did fail
because of the URL length and would have been extremely slow.
Now the search request is done in a POST and the input is moved to a
request body.
The parsing to a `SearchExpression` is correct now as well and no longer
has a backtrack explosion.
Then a new syntax becomes introduced:
```
hdl:"11159/632043,11159/632043"
```
This will search for multiple handles and does it fast by no longer
doing a fulltext based search. Instead it's exact match.
Caveat: It does not support wildcards yet, because it would slow down
everything by factor 10.
Other keywords supporting this syntax follow if needed.