fix(706): search filter and search expression for zdb/sig now distinct
The filters for zdb and paket sigels work completely differently when
used in a search expression and in the search bar. Before this commit
both usages would result in the same query which was wrong. A zdb key
with multiple words is always an OR operation while the bad always is an
AND operation.